+++ Datadog/IntegrationUnitTests/RUM/UIApplicationSwizzlerTests.swift
@@ -9,7 +9,7 @@ import TestUtilities
 @testable import DatadogRUM
 
 // TODO: RUMM-2034 Remove this flag once we have a host application for tests
-#if !os(tvOS)
+#if os(iOS)
 
 class UIApplicationSwizzlerTests: XCTestCase {
     private let handler = RUMActionsHandlerMock()