+++ DatadogCore/Tests/Datadog/DatadogCore/Context/NetworkConnectionInfoPublisherTests.swift
@@ -6,7 +6,6 @@
 
 import XCTest
 import Network
-import SystemConfiguration
 import DatadogInternal
 @testable import DatadogCore
 
@@ -48,7 +47,7 @@ class NetworkConnectionInfoConversionTests: XCTestCase {
 
     #if compiler(>=6.2)
     func testNWPathLinkQuality() throws {
-        guard #available(iOS 26.0, tvOS 26.0, macOS 26.0, watchOS 26.0, *) else {
+        guard #available(iOS 26.0, tvOS 26.0, macOS 26.0, watchOS 26.0, visionOS 26.0, *) else {
             throw XCTSkip("NWPath.LinkQuality requires iOS 26+")
         }
         XCTAssertEqual(LinkQuality(.good), .good)