+++ DatadogRUM/Sources/Instrumentation/Views/UIKit/UIViewControllerSwizzler.swift
@@ -4,6 +4,7 @@
  * Copyright 2019-Present Datadog, Inc.
  */
 
+#if !os(watchOS)
 import UIKit
 import DatadogInternal
 
@@ -78,3 +79,4 @@ internal class UIViewControllerSwizzler {
         }
     }
 }
+#endif